+++ lib/src/foundation/feature_discovery.dart
@@ -16,26 +16,31 @@ enum ContentLocation {
 }
 
 class FeatureDiscovery extends StatelessWidget {
+  static Bloc _blocOf(BuildContext context) {
+    try {
+      return Bloc.of(context);
+    } on BlocNotFoundError catch (e) {
+      throw FlutterError(e.message +
+          '\nEnsure that it also wraps the context of the ${context.widget.runtimeType} widget from which you have called a static method in FeatureDiscovery.');
+    }
+  }
+
   /// Steps are the featureIds of the overlays.
   /// Though they can be placed in any [Iterable], it is recommended to pass them as a [Set]
   /// because this ensures that every step is only shown once.
   static void discoverFeatures(BuildContext context, Iterable<String> steps) =>
-      Bloc.of(context).discoverFeatures(steps: steps.toList());
+      _blocOf(context).discoverFeatures(steps: steps.toList());
 
   /// This will schedule completion of the current discovery step and continue
   /// onto the step after the completion animation of the current overlay if successful.
-  ///
-  /// The [stepId] ensures that you are marking the correct feature for completion.
-  /// If the provided [stepId] does not match the feature that is currently shown, i.e.
-  /// the currently active step, nothing will happen.
   static void completeCurrentStep(BuildContext context) =>
-      Bloc.of(context).completeStep();
+      _blocOf(context).completeStep();
 
   /// This will schedule dismissal of the current discovery step and with that
   /// of the current feature discovery. The dismissal animation will play if successful.
   /// If you want to complete the step and continue the feature discovery,
   /// call [completeCurrentStep] instead.
-  static void dismissAll(BuildContext context) => Bloc.of(context).dismiss();
+  static void dismissAll(BuildContext context) => _blocOf(context).dismiss();
 
   @Deprecated("Use [dismissAll] instead.")
   static void dismiss(BuildContext context) => dismissAll(context);
@@ -44,11 +49,11 @@ class FeatureDiscovery extends StatelessWidget {
   /// of the [DescribedFeatureOverlay] that is currently supposed to be shown, or `null`.
   ///
   /// Note that this will also return the feature id of the current step of the steps
-  /// you passed to [discoverFeature] even when there is no [DescribedFeatureOverlay]
+  /// you passed to [discoverFeatures] even when there is no [DescribedFeatureOverlay]
   /// in the tree to display the overlay.
   /// This means that you cannot use this to check if a feature overlay is being displayed.
   static String currentFeatureIdOf(BuildContext context) =>
-      Bloc.of(context).activeFeatureId;
+      _blocOf(context).activeFeatureId;
 
   @Deprecated("Use [currentFeatureIdOf] instead.")
   static String activeFeatureId(BuildContext context) =>